requirements.txt
beautifulsoup4
numpy
pandas
requests
lxml
openai
flask
pyyaml